src/main/java/ua/com/juja/sqlcmd/controller/command/Drop.java
package ua.com.juja.sqlcmd.controller.command;

import ua.com.juja.sqlcmd.model.DatabaseManager;
import ua.com.juja.sqlcmd.view.View;

public class Drop implements Command {
    private final DatabaseManager manager;
    private final View view;

    public Drop(DatabaseManager manager, View view) {
        this.manager = manager;
        this.view = view;
    }

    @Override
    public boolean canProcess(String command) {
        return command.startsWith("drop|");
    }

    @Override
    public void process(String command) {
        String[] data = command.split("\\|");
        if (data.length != 2) {
            throw new IllegalArgumentException("command format 'drop|tableName' but you entered: " + command);
        }
        String tableName = data[1];

        manager.drop(tableName);
        view.write(String.format("the table '%s' was deleted successfully", tableName));
    }
}
